pkg/status: move under pkg/vmextension
Logically makes sense to move pkg/status under pkg/vmextension/status
so that other extension handlers importing these utilities will just
be vendoring pkg/vmextension/*.
